+++ packages/node-vibrant/src/browser.ts
@@ -1,6 +1,6 @@
-import Vibrant from "./configs/browser";
-import pipeline from "./pipeline";
+import { Vibrant } from "./configs/browser";
+import { pipeline } from "./pipeline";
 
 Vibrant.use(pipeline);
 
-export default Vibrant;
+export { Vibrant };