+++ packages/node-vibrant/src/configs/browser.ts
@@ -1,6 +1,6 @@
-import Vibrant from "./config";
-import BrowserImage from "@vibrant/image-browser";
+import { BrowserImage } from "@vibrant/image-browser";
+import { Vibrant } from "./config";
 
 Vibrant.DefaultOpts.ImageClass = BrowserImage;
 
-export default Vibrant;
+export { Vibrant };