+++ packages/node-vibrant/src/worker.ts
@@ -1,8 +1,8 @@
-import Vibrant from "./configs/browser";
 import { WorkerPipeline } from "@vibrant/core";
+import { Vibrant } from "./configs/browser";
 
 import PipelineWorker from "./pipeline/index.worker?worker";
 
 Vibrant.use(new WorkerPipeline(PipelineWorker as never));
 
-export default Vibrant;
+export { Vibrant };