+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## Breaking Changes
 
 - You now must import from `node-vibrant/browser`, `node-vibrant/node`, or `node-vibrant/worker` to get the correct environment-specific implementation
-- `Vibrant` class is now a default export
+- `Vibrant` class is now a named export
 - Node 18+ is now required
 - ES5 support is dropped